Fix issues from static analysis (#202)
* Casting a signed value which has type 'int' to a bigger unsigned integer type 'size_t' while initializing a variable
* Assignment of a signed value which has type 'int32_t' to a variable of a bigger integer type 'size_t'
* Uninitialized field

+++ shell/platform/tizen/flutter_tizen_engine.cc
@@ -348,8 +348,8 @@ void FlutterTizenEngine::SendWindowMetrics(int32_t width,
                                            double pixel_ratio) {
   FlutterWindowMetricsEvent event = {};
   event.struct_size = sizeof(FlutterWindowMetricsEvent);
-  event.width = width;
-  event.height = height;
+  event.width = static_cast<size_t>(width);
+  event.height = static_cast<size_t>(height);
   if (pixel_ratio == 0.0) {
     // The scale factor is computed based on the display DPI and the current
     // profile. A fixed DPI value (72) is used on TVs. See:
